Massive Vue Js Update hope it works on beta....
