+++ Wargame/TMCW/Collapse/MultiStepUnit.php
@@ -95,15 +95,8 @@ class MultiStepUnit extends \Wargame\MovableUnit  implements \JsonSerializable
         }
         $strength = $this->getUnmodifiedStrength();
         if($this->id == 54){
-//            echo $strength;
         }
         $strength = $this->getCombatAdjustments($strength);
-        if($this->id == 54){
-//            echo " $strength";
-        }
-//        if($strength < 0.1){
-//            $strength = .1;
-//        }
 
         return $strength;
     }
