+++ Wargame/TMCW/NorthVsSouth/MultiStepUnit.php
@@ -94,16 +94,8 @@ class MultiStepUnit extends \Wargame\MovableUnit  implements \JsonSerializable
             return $this->origSteps > $this->steps;
         }
         $strength = $this->getUnmodifiedStrength();
-        if($this->id == 54){
-//            echo $strength;
-        }
+
         $strength = $this->getCombatAdjustments($strength);
-        if($this->id == 54){
-//            echo " $strength";
-        }
-//        if($strength < 0.1){
-//            $strength = .1;
-//        }
 
         return $strength;
     }
