+++ Wargame/Vu/vu.scss
@@ -29,7 +29,7 @@
     font-weight: 400;
   }
 }
-.blueUnit, .blue, .playerBlue {
+.blueUnit, .blue, .playerBlue , .header-Blue{
   background-color: rgb(132, 181, 255);
   border-color: rgb(132, 181, 255);
 }
@@ -69,7 +69,7 @@
   background-color: rgb(223, 88, 66);
   border-color: rgb(223, 88, 66);
 }
-.brightRed, .british, .playerRed {
+.brightRed, .british, .playerRed, .header-Red {
   background-color: rgb(223, 88, 66);
   border-color: rgb(223, 88, 66);
 }
