+++ Wargame/wargame-helpers/Vue/scss/vue-misc.scss
@@ -1,0 +1,64 @@
+body{
+  overflow:hidden;
+}
+.clear{
+  clear:both;
+}
+.btn-group button{
+  font-size: 20px;
+}
+
+.info-dropdown{
+  li{
+    font-weight: 400;
+  }
+}
+
+#vue-crt {
+  display:none;
+}
+
+#crt-wrapper.open  #vue-crt{
+  display: block;
+}
+
+#crt-wrapper {
+}
+
+.handle{
+  display:none !important;
+}
+
+
+.highlighted{
+  background: rgba(255,255,0,.6);
+}
+
+.row-hex{
+  width:71px;
+}
+.mapFlashSymbols{
+  position:absolute;
+  color:white;
+  background-color: transparent;
+  margin-top:-31px;
+  margin-left:-36px;
+}
+
+.flashMessage {
+  position: fixed;
+  width: 80%;
+  background: rgba(255, 255, 255, .9);
+  text-align: center;
+  border-radius: 30px;
+  border: 10px solid black;
+  font-size: 100px;
+  z-index: 1000;
+  pointer-events: none;
+}
+
+#floatMessageContainer{
+  position: absolute;
+  top:200px;
+  left: 100px;
+}
