+++ Wargame/wargame-helpers/Vue/syncObj.js
@@ -1,0 +1,3 @@
+import {Sync} from "../Sync";
+export const syncObj = new Sync(fetchUrl);
+DR.sync = syncObj;
